+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # Changelog
 
+## 0.1.1
+ - Added keybind for toggling the UINetwork debug ui per request - https://github.com/darmuh/YAPYAP_LocalMultiplayer/issues/1
+	- Keybind is in config section ``Settings`` under ``Toggle NetworkDebug UI``
+
 ## 0.1.0
-- Initial release.
+ - Initial release.